requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
captcha==0.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.35.0
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
np-utils==0.6.0
numpy==2.1.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
Pygments==2.19.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.4
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.107
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.19.0
